+++ src/activities/ActivityResult.h
@@ -91,10 +91,6 @@ struct ReadingStatsResult {
   bool changed = false;
 };
 
-struct Ao3InfoResult {
-  bool checkUpdates = false;
-};
-
 struct ClippingResult {
   std::string text;
   int fromWordIdx = -1;
@@ -126,8 +122,7 @@ struct ClippingJumpResult {
 using ResultVariant = std::variant<std::monostate, WifiResult, KeyboardResult, MenuResult, ChapterResult, PercentResult,
                                    IntervalResult, OptionSelectionResult, PageResult, ProgressChangeResult, SyncResult,
                                    NetworkModeResult, FootnoteResult, BookmarkResult, FileBrowserActionResult,
-                                   FilePathResult, WordResult, ReadingStatsResult, Ao3InfoResult, ClippingResult,
-                                   ClippingJumpResult>;
+                                   FilePathResult, WordResult, ReadingStatsResult, ClippingResult, ClippingJumpResult>;
 
 struct ActivityResult {
   bool isCancelled = false;